training: add single-GPU smoke test + fix bf16-native RecurrentBlock dtype
Adds training/smoke_train.py: a short single-GPU smoke test that streams real
FineWeb-Edu data through mythos_1b end to end (tokenizer → dataset → fwd/bwd →
AdamW step → checkpoint save+reload), reusing the helpers from the cluster
script 3b_fine_web_edu.py. mythos_1b with the ~200k gpt-oss vocab is ~1.41B
params; fp32 params+grads+Adam is ~22.5 GB and OOMs a 24 GB card, so the smoke
test runs bf16-native (~11.3 GB, peak ~18.5 GB observed).

Running bf16-native surfaced a dtype bug in RecurrentBlock.forward: cumulative_p
was initialized as float32 (no dtype=) and still_running was cast with .float(),
which promoted the ACT-weighted h_out to float32 and collided with the bf16
Coda weights downstream. Initialize cumulative_p with dtype=h.dtype and cast
still_running with .to(h.dtype). Behavior is identical in fp32 (existing 76
tests still pass); this only enables low-precision single-GPU training without
autocast.

Also gitignore checkpoint outputs (checkpoints/, checkpoints_smoke/, *.pt).

+++ open_mythos/main.py
@@ -855,7 +855,11 @@ class RecurrentBlock(nn.Module):
         B, T, D = h.shape
 
         halted = torch.zeros(B, T, device=h.device, dtype=torch.bool)
-        cumulative_p = torch.zeros(B, T, device=h.device)
+        # Match h's dtype so the ACT-weighted accumulation below stays in the
+        # model's compute dtype. Defaulting to float32 here would promote h_out
+        # to float32 and collide with bf16/fp16 weights in the downstream Coda
+        # when training in low precision without autocast.
+        cumulative_p = torch.zeros(B, T, device=h.device, dtype=h.dtype)
         h_out = torch.zeros_like(h)
 
         for t in range(n_loops):
@@ -880,10 +884,10 @@ class RecurrentBlock(nn.Module):
                 remainder,
                 p,
             )
-            weight = weight * still_running.float()
+            weight = weight * still_running.to(h.dtype)
             h_out = h_out + weight.unsqueeze(-1) * h
 
-            cumulative_p = cumulative_p + p * still_running.float()
+            cumulative_p = cumulative_p + p * still_running.to(h.dtype)
             halted = halted | (cumulative_p >= self.cfg.act_threshold)
 
             # Only short-circuit when there is no KV cache to keep consistent.
